Fix flapping tests
By removing occurrences of `sleep 0.1`. The difficulty here is to make
tests fail if a message is not delivered (instead of hanging
forever). This also makes tests significantly faster (because tests
can immediately proceed instead of having to wait for a fixed amount
of time).

There are still some `sleep` calls left, but these are with much
larger timeouts (1 or 2 seconds), so it should be fine for now.

Resolves #28.